requirements.txt
django
pandas
nsepy
psycopg2
djangorestframework
lxml
gunicorn